remove Table, Group, ComponentContainer. Replace with HTML equivalents
