+++ src/components/index.ts
@@ -10,9 +10,6 @@ export * from "./gridlines";
 export * from "./group";
 export * from "./guideLineLayer";
 export * from "./htmlComponent";
-export * from "./htmlComponentContainer";
-export * from "./htmlGroup";
-export * from "./htmlTable";
 export * from "./interpolatedColorLegend";
 export * from "./label";
 export * from "./legend";
